refactor(metax): isolate DSV4 INT8 experts
